ANLP_2_multi-agent-assistant/requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
jupyter>=1.0.0

